Please refer to the 1.1 update notes

+++ Sources/SuiKit/Protocols/Sui/TypeProtocol.swift
@@ -28,5 +28,5 @@ import Foundation
 public protocol TypeProtocol: KeyProtocol {
     /// Returns the type variant the class represents
     /// - Returns: An Integer value that represents the class's type tag
-    func variant() -> Int
+    func variant() -> UInt8
 }
